+++ src/app/(main)/_components/SectionMoiveList/MovieTopRank/Mobile_BestMovie.tsx
@@ -115,7 +115,7 @@ export default function Mobile_BestMovie(MovieData: Mobile_BestMoiveProps) {
                               alt="white_ start"
                               className="h-4 w-4"
                             />
-                            {MovieDetailData.reviewList[0].likeCount && (
+                            {MovieDetailData.reviewList[0]?.likeCount && (
                               <span>
                                 {MovieDetailData.reviewList[0].likeCount}
                               </span>
@@ -123,11 +123,11 @@ export default function Mobile_BestMovie(MovieData: Mobile_BestMoiveProps) {
                           </div>
                         </div>
                         <div className="flex h-[42px] items-center gap-2">
-                          {MovieDetailData?.reviewList[0].profile && (
+                          {MovieDetailData?.reviewList[0]?.profile && (
                             <div
                               className="h-[24px] w-[24px] rounded-[60px] "
                               style={{
-                                backgroundImage: `linear-gradient(180deg, rgba(0, 0, 0, 0.0) 0%, rgba(0, 0, 0, 0) 100%), url(data:image/jpeg;base64,${MovieDetailData?.reviewList[0].profile}`,
+                                backgroundImage: `linear-gradient(180deg, rgba(0, 0, 0, 0.0) 0%, rgba(0, 0, 0, 0) 100%), url(data:image/jpeg;base64,${MovieDetailData?.reviewList[0]?.profile}`,
                                 backgroundSize: "cover",
                                 backgroundPosition: "center",
                               }}
@@ -135,7 +135,7 @@ export default function Mobile_BestMovie(MovieData: Mobile_BestMoiveProps) {
                           )}
 
                           <span className="line-clamp-2 w-[174px] text-Silver Text-s-Regular">
-                            {MovieDetailData.reviewList[0].content &&
+                            {MovieDetailData.reviewList[0]?.content &&
                               MovieDetailData.reviewList[0].content}
                           </span>
                         </div>
