+++ src/app/(main)/_components/SectionMoiveList/WatchedMovies/WatchedMoive.tsx
@@ -1,26 +1,54 @@
+"use client";
 import { Suspense, useEffect, useState } from "react";
 
 import SpeechBubble from "@/components/speechBubble/SpeechBubble";
 import { tokenManager } from "@/services/auth/tokenManager";
+import { movieAPIs } from "@/services/movie/movieAPIs";
 import { myAPIs } from "@/services/my/myAPIs";
 
+import WatchedSkeleton from "../../MainSkeleton/WatchedMoive/WatchedSkeleton";
 // import WatchedSkeleton from "../../MainSkeleton/WatchedMoive/WatchedSkeleton";
 import MoviePosts from "./MoviePosts";
 
-export default async function WatchedMoive() {
-  const data = await myAPIs.getUser();
-  let message = data
-    ? "톡을 많이 작성할수록 내 취향에 비슷해져요"
-    : "로그인 하고 별을 눌러 평가해보세요 ";
+export default function WatchedMoive() {
+  const [accsstoken, setaccsstoken] = useState<string | null>("");
 
-  return (
+  const [WatchMovieData, setWatchMovieData] = useState<WatchMovie[]>([]);
+
+  const [isLoading, setIsLoading] = useState(true);
+
+  useEffect(() => {
+    const fetchMovie = async () => {
+      try {
+        const response: WatchMovie[] = await movieAPIs.getWatchMovie();
+        setWatchMovieData(response);
+        const token = tokenManager.getToken();
+        if (token) {
+          setaccsstoken(token);
+        }
+      } catch (error) {
+        console.error("Failed to fetch movie data", error);
+      } finally {
+        setIsLoading(false);
+      }
+    };
+
+    fetchMovie();
+  }, []);
+  let message =
+    accsstoken == ""
+      ? "톡을 많이 작성할수록 내 취향에 비슷해져요"
+      : "로그인 하고 별을 눌러 평가해보세요 ";
+  return isLoading ? (
+    <WatchedSkeleton />
+  ) : WatchMovieData.length != 0 ? (
     <div className="flex flex-col gap-5">
       <div className="flex flex-col gap-1 text-center  Laptop:flex-row Laptop:gap-5 Laptop:text-left">
         <h1 className=" text-Silver Text-l-Bold  Laptop:Text-xxl-Bold">
           혹시 이 영화 보셨나요?
         </h1>
         <div className="hidden  Laptop:block">
-          {data ? (
+          {accsstoken ? (
             <SpeechBubble id={"WatchedMovie"} dir="left">
               {message}
             </SpeechBubble>
@@ -35,8 +63,8 @@ export default async function WatchedMoive() {
         </span>
       </div>
       {/* <Suspense fallback={<WatchedSkeleton />}> */}
-      <MoviePosts />
+      <MoviePosts WatchMovieData={WatchMovieData} />
       {/* </Suspense> */}
     </div>
-  );
+  ) : null;
 }
